Extract FeedbackSection as component
